Fix: (ts-human-format-duration) Previous fix
The fix applied previously was testing the value of SECONDS after it
was bound to the plist' SECONDS :value, which can be 0 even if the
SECONDS passed to the function is non-zero.  The tests added didn't
catch this, because they tested with values that ts-human-duration
returned 0 :SECONDS for.  Oops.

Released as 0.2.2.